rename UnwrapFutureCompatibleIfRector to UnwrapFutureCompatibleIfFunctionExistsRector
